In this pull request, I've extended the `joinName` helper to make it work with "complex" field names. There are a lot of possible names that are impossible for uniforms at the moment - most importantly the ones with dots in names (see #963).

Most importantly, this change is backward compatible for all schemas with alphanumeric fields only (I believe it covers more than 99% of projects). When it comes to the rest, like the ones with fields called `/` or `[`, it may not be the case (it probably will with most of them; I still have to test it).

The point of this change is that `joinName` correctly understands JSON Path-like names. The rules are simple and very similar to the ones in JavaScript:
* `a` works as before, accessing top-level fields.
* `a.b` works as before, accessing objects' properties.
* `a.0` works as before, accessing array elements.
* `["a"]` is a new way of accessing top-level fields.
* `["a"]["b"]` is a new way of accessing objects' properties.
  * You can mix them freely: `a["b"].c` is the same as `a.b.c` or `["a"].b["c"]`.
  * You can use all of the characters within the subscript (`[""]`) with only one limitation: `"` has to be escaped into `\"`.
    * If you'd like to render field called `[""]`, then `["[\"\"]"]` it is.

There are however three things that I'm not entirely happy with:
1. **The fact that `joinName` _always_ returns _something_.** This doesn't sound like a problem, but it is - there are some potentially incorrect names, that we handle in a given way (based on the current implementation). For example `[` is normalized to `["["]` and `['']` is normalized to `["['']"]`. See the [tests](https://github.com/vazco/uniforms/blob/69f2e958d4f96123be8b4a8773d1f8f9ab98f5b8/packages/uniforms/__tests__/joinName.ts#L86-L110) for more examples. (Maybe we should add a development-only warning?)
2. **More semi-public API.** As the escaping is needed for the forms to work, the bridge that would like to support "complex" names should use `joinName.escape` and `joinName.unescape` functions. I've marked them as `@internal`, but they are still visible and typed.

---

Performance changes.

Example data taken from the `joinName` test cases: less than 1% slower on average.

|        |   min  |   avg  |   max  |
|:------:|:------:|:------:|:------:|
| Before |  912ms |  936ms |  982ms |
| After  |  925ms |  936ms |  991ms |

Example data taken from an actual project: 7% slower on average.

|        |   min  |   avg  |   max  |
|:------:|:------:|:------:|:------:|
| Before |  220ms |  228ms |  239ms |
| After  |  239ms |  244ms |  259ms |

---

<details>
<summary>Here's an example schema so you can test it yourself.</summary>

```js
(() => {
  const ajv = new Ajv({ allErrors: true, useDefaults: true, keywords: ["uniforms"] });
  const schema = {
    type: 'object',
    properties: {
      complexNames: {
        type: 'object',
        properties: {
          a: {
            type: 'array',
            items: {
              type: 'object',
              properties: {
                b: {
                  type: 'object',
                  properties: {
                    'c-d': {
                      type: 'array',
                      items: {
                        type: 'object',
                        properties: {
                          "f/'g": { type: 'string', pattern: '[0-9]{5}' },
                          'h/"i': { type: 'string', pattern: '[0-9]{5}' },
                          'j\'/"k': { type: 'string', pattern: '[0-9]{5}' },
                        },
                      },
                    },
                  },
                },
              },
            },
          },
        },
      },
      'path.with.a.dot': {
        type: 'object',
        properties: {
          'another.with.a.dot': {
            type: 'string',
          },
          another: {
            type: 'object',
            properties: {
              with: {
                type: 'object',
                properties: {
                  a: {
                    type: 'object',
                    properties: {
                      dot: {
                        type: 'number',
                      },
                    },
                  },
                },
              },
            },
          },
        },
      },
      path: {
        type: 'object',
        properties: {
          'with.a.dot': { type: 'number' },
          '["with.a.dot"]': { type: 'string' },
        },
      },
    },
  };

  function createValidator(schema) {
    const validator = ajv.compile(schema);

    return (model) => {
      validator(model);

      if (validator.errors && validator.errors.length) {
        return { details: validator.errors };
      }
    };
  }

  const schemaValidator = createValidator(schema);

  return new JSONSchemaBridge(schema, schemaValidator);
})()
```

</details>

I took the tests from #1078 but went with a completely different approach.